+++ home.mmgen/scripts/alarm_clock.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+eval 'btc-ticker --resize-window --alarm-clock-only'