+++ upgrade/0.0.7/upgrade-f.sh
@@ -2,12 +2,14 @@
 #:REV: 0.0.7f
 #:DESC: install unzip and pycurl packages
 #:DESC: upgrade MMGen to commit eb6f7ef195bd391174bd6563d0c85b1f1934873f
-#:DESC: this rev enables installation of node-tools
+#:DESC: install node-tools
 
 set -e
 
 function install_unzip_pycurl {
 	echo "Installing packages 'unzip' and 'python-pycurl'"
+	eval "$APT_GET update"
+	eval "$APT_GET upgrade"
 	eval "$APT_GET install unzip python-pycurl"
 }
 
@@ -19,9 +21,19 @@ function upgrade_mmgen {
 	)
 }
 
+function install_node_tools {
+	(
+		echo "Installing mmgen-node-tools"
+		cd /setup/git/node-tools
+		sudo python ./setup.py install
+	)
+}
+
 which unzip >/dev/null && dpkg -l python-pycurl >/dev/null 2>&1 || install_unzip_pycurl
 
 COMMIT='eb6f7ef195bd391174bd6563d0c85b1f1934873f'
 [ -d ~mmgen/src/mmgen-$COMMIT ] || upgrade_mmgen
+
+which btc-ticker >/dev/null || install_node_tools
 
 exit 0